thumbnailer/pyproject.toml
[project]
authors = [
    {name = "Mathieu Leplatre", email = "mathieu@leplat.re"},
]
requires-python = "<4.0,>=3.11"
dependencies = [
    "inotify-simple>=1.3.5,<3.0.0",
    "watchdog>=4.0.1,<7.0.0",
    "pillow>=10.3.0,<13.0.0",
    "moviepy>=2.2.1,<3.0.0",
    "ffmpeg-python<1.0.0,>=0.2.0",
    "pymupdf<2.0.0,>=1.24.5",
    "pillow-heif>=0.16.0,<2.0.0",
]
name = "thumbnailer"
version = "0.1.0"
description = ""
readme = "README.md"

[dependency-groups]
dev = [
    "numpy>=2.3.2",
    "pytest>=8.4.1",
    "pytest-coverage>=0.0",
    "ruff>=0.12.9",
]
